didact.cpp: split out of Candela computers in its own driver for clarity
